+++ src/OrchardCore.Modules/OrchardCore.Users/Services/UserEmailService.cs
@@ -1,0 +1,103 @@
+using System.Text;
+using System.Text.Encodings.Web;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.Localization;
+using OrchardCore.DisplayManagement;
+using OrchardCore.Email;
+using OrchardCore.Mvc.Core.Utilities;
+using OrchardCore.Users.Controllers;
+using OrchardCore.Users.Models;
+using OrchardCore.Users.ViewModels;
+namespace OrchardCore.Users.Services;
+
+public sealed class UserEmailService
+{
+    private readonly IHttpContextAccessor _httpContextAccessor;
+    private readonly LinkGenerator _linkGenerator;
+    private readonly UserManager<IUser> _userManager;
+    private readonly IEmailService _emailService;
+    private readonly IDisplayHelper _displayHelper;
+    private readonly HtmlEncoder _htmlEncoder;
+
+    internal readonly IStringLocalizer S;
+
+    public UserEmailService(
+        IHttpContextAccessor httpContextAccessor,
+        LinkGenerator linkGenerator,
+        UserManager<IUser> userManager,
+        IEmailService emailService,
+        IDisplayHelper displayHelper,
+        HtmlEncoder htmlEncoder,
+        IStringLocalizer<UserEmailService> stringLocalizer)
+    {
+        _httpContextAccessor = httpContextAccessor;
+        _linkGenerator = linkGenerator;
+        _userManager = userManager;
+        _emailService = emailService;
+        _displayHelper = displayHelper;
+        _htmlEncoder = htmlEncoder;
+        S = stringLocalizer;
+    }
+
+    public async Task<bool> SendEmailConfirmationAsync(IUser user)
+    {
+        var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+
+        var callbackUrl = _linkGenerator.GetUriByAction(_httpContextAccessor.HttpContext, nameof(EmailConfirmationController.ConfirmEmail), typeof(EmailConfirmationController).ControllerName(),
+        new
+        {
+            userId = await _userManager.GetUserIdAsync(user),
+            code,
+        });
+
+        var email = await _userManager.GetEmailAsync(user);
+
+        if (string.IsNullOrEmpty(email))
+        {
+            return false;
+        }
+
+        return await SendEmailAsync(email, S["Confirm your account"], new ConfirmEmailViewModel
+        {
+            User = user,
+            ConfirmEmailUrl = callbackUrl,
+        });
+    }
+
+    public async Task<bool> SendPasswordResetAsync(User user)
+    {
+        user.ResetToken = Convert.ToBase64String(Encoding.UTF8.GetBytes(user.ResetToken));
+
+        var resetPasswordUrl = _linkGenerator.GetUriByAction(_httpContextAccessor.HttpContext, nameof(ResetPasswordController.ResetPassword), typeof(ResetPasswordController).ControllerName(),
+        new
+        {
+            userId = await _userManager.GetUserIdAsync(user),
+            code = user.ResetToken,
+        });
+
+        // send email with callback link
+        return await SendEmailAsync(user.Email, S["Reset your password"], new LostPasswordViewModel()
+        {
+            User = user,
+            LostPasswordUrl = resetPasswordUrl
+        });
+    }
+
+    private async Task<bool> SendEmailAsync(string email, string subject, IShape model)
+    {
+        var body = string.Empty;
+
+        using (var sw = new StringWriter())
+        {
+            var htmlContent = await _displayHelper.ShapeExecuteAsync(model);
+            htmlContent.WriteTo(sw, _htmlEncoder);
+            body = sw.ToString();
+        }
+
+        var result = await _emailService.SendAsync(email, subject, body);
+
+        return result.Succeeded;
+    }
+}
